+++ analysis/report/plot_measures_report.py
@@ -5,8 +5,6 @@ import pathlib
 import matplotlib.pyplot as plt
 import seaborn as sns
 from report_utils import (
-    ci_to_str,
-    ci_95_proportion,
     add_date_lines,
     autoselect_labels,
     colour_palette,
@@ -14,6 +12,8 @@ from report_utils import (
     parse_date,
     coerce_numeric,
     set_fontsize,
+    make_season_table,
+    annotate_seasons,
     MEDICATION_TO_CODELIST,
     CLINICAL_TO_CODELIST,
 )
@@ -21,87 +21,6 @@ from report_utils import (
 import matplotlib.ticker as ticker
 
 ticker.Locator.MAXTICKS = 10000
-
-
-# NOTE: bug with pandas 1.01, cannot do pivot with MultiIndex
-def MultiIndex_pivot(
-    df: pd.DataFrame,
-    index: str = None,
-    columns: str = None,
-    values: str = None,
-) -> pd.DataFrame:
-    """
-    https://github.com/pandas-dev/pandas/issues/23955
-    Usage:
-    df.pipe(MultiIndex_pivot, index = ['idx_column1', 'idx_column2'],
-            columns = ['col_column1', 'col_column2'], values = 'bar')
-    """
-    output_df = df.copy(deep=True)
-    if index is None:
-        names = list(output_df.index.names)
-        output_df.reset_index(drop=True, inplace=True)
-    else:
-        names = index
-    output_df = output_df.assign(
-        tuples_index=[tuple(i) for i in output_df[names].values]
-    )
-    if isinstance(columns, list):
-        output_df = output_df.assign(
-            tuples_columns=[tuple(i) for i in output_df[columns].values]
-        )
-        output_df = output_df.pivot(
-            index="tuples_index", columns="tuples_columns", values=values
-        )
-        output_df.columns = pd.MultiIndex.from_tuples(
-            [((x[0],) + x[1]) for x in output_df.columns],
-            names=[None] + columns,
-        )
-    else:
-        output_df = output_df.pivot(
-            index="tuples_index", columns=columns, values=values
-        )
-    output_df.index = pd.MultiIndex.from_tuples(output_df.index, names=names)
-    return output_df
-
-
-def produce_min_max_table(df, column_to_plot, category):
-    df.numerator = df.numerator.astype(float)
-    df.denominator = df.denominator.astype(float)
-    cis = ci_95_proportion(df, scale=1000)
-    df["Rate (95% CI)"] = ci_to_str(cis)
-    df = df.rename({"numerator": "Count"}, axis=1)
-    table = df.pipe(
-        MultiIndex_pivot,
-        index=[category, "type"],
-        columns="gas_year",
-        values=["Count", "rate", "Rate (95% CI)"],
-    )
-    if column_to_plot == "numerator":
-        table["Count"] = table["Count"].applymap(lambda x: f"{x:.0f}")
-        return table["Count"]
-    else:
-        table.Count = table.Count.astype(float)
-        table.rate = table.rate.astype(float)
-        ratio = table["rate"]["2023"] / table["rate"]["2018"]
-
-        # See formula of ci irr:
-        # https://researchonline.lshtm.ac.uk/id/eprint/251164/1/pmed.1001270.s005.pdf
-        sd_log_irr = np.sqrt(
-            (1 / table["Count"]["2023"]) + (1 / table["Count"]["2018"])
-        )
-        lci = np.exp(np.log(ratio) - 1.96 * sd_log_irr)
-        uci = np.exp(np.log(ratio) + 1.96 * sd_log_irr)
-        rr_cis = pd.concat([ratio, lci, uci], axis=1)
-
-        rr_cis_str = ci_to_str(rr_cis)
-        rr_cis_str.name = ("2023 v 2018", "Rate Ratio (95% CI)")
-
-        # Create table with count, rate (95% CI)
-        table["Count"] = table["Count"].applymap(lambda x: f"{x:.0f}")
-        count_cis = (table[["Count", "Rate (95% CI)"]]).swaplevel(axis=1)
-        cols = count_cis.columns.get_level_values(0).unique()
-        reordered = count_cis.reindex(columns=cols, level="gas_year")
-        return pd.concat([reordered, rr_cis_str], axis=1)
 
 
 def plot_measures(
@@ -177,53 +96,14 @@ def plot_measures(
         ax.set_xticks(xticks)
         ax.set_xticklabels([x.strftime("%B %Y") for x in xticks])
 
+        # TODO: check that category exists?
         if produce_season_table or mark_seasons:
-            # TODO: check whether this works for weekly
-            df_copy["gas_year"] = pd.to_datetime(df_copy.index).to_period(
-                "A-Aug"
+            season_table = make_season_table(
+                df_copy, category, column_to_plot, output_dir, filename
             )
-            mins_and_maxes = []
-            for year, data in df_copy.groupby(["gas_year", category]):
-                data_sorted = data.dropna(subset=[column_to_plot]).sort_values(
-                    column_to_plot
-                )
-                if data_sorted.empty:
-                    continue
-                data_min = data_sorted.iloc[0]
-                data_max = data_sorted.iloc[-1]
-                if mark_seasons:
-                    plt.annotate(
-                        int(data_min[column_to_plot]),
-                        xy=(data_min.name, data_min[column_to_plot]),
-                        xytext=(0, 25),
-                        textcoords="offset points",
-                        verticalalignment="bottom",
-                        arrowprops=dict(facecolor="black", shrink=0.025),
-                        horizontalalignment="left"
-                        if (data_min.name.month % 2) == 1
-                        else "right",
-                        fontsize=12,
-                    )
-                    plt.annotate(
-                        int(data_max[column_to_plot]),
-                        xy=(data_max.name, data_max[column_to_plot]),
-                        xytext=(0, -25),
-                        textcoords="offset points",
-                        verticalalignment="top",
-                        arrowprops=dict(facecolor="blue", shrink=0.025),
-                        fontsize=12,
-                    )
-                data_min["type"] = "min"
-                data_max["type"] = "max"
-                mins_and_maxes.append(data_min)
-                mins_and_maxes.append(data_max)
-            if produce_season_table:
-                table = produce_min_max_table(
-                    pd.concat(mins_and_maxes, axis=1).T,
-                    column_to_plot,
-                    category,
-                )
-                table.to_html(output_dir / f"{filename}_table.html")
+            # TODO: check whether this works for weekly
+            if mark_seasons:
+                annotate_seasons(season_table, column_to_plot, ax)
 
     elif frequency == "week":
         xticks = pd.date_range(
